agents/requirements.txt
google-adk==1.5.0
litellm>=1.73.6
openai>=1.68.2
pydantic==2.11.1
fastapi==0.118.3
starlette==0.48.0
click==8.3.0
google-api-python-client==2.186.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.95.1
google-cloud-storage==2.18.0
google-genai==1.27.0
